requirements.txt
Flask==2.0.1
pika==1.2.0
requests==2.25.1
python-dotenv==0.19.1
gunicorn==20.1.0
